+++ src/commands/showRunningMacros.ts
@@ -18,7 +18,10 @@ export async function showRunningMacros(manager: Manager) {
     return;
   }
 
-  const selected = await pickMacroFile({ '': runningItems.map((item) => item.runInfo.macro.uri) });
+  const selected = await pickMacroFile(
+    { '': runningItems.map((item) => item.runInfo.macro.uri) },
+    { hideOpen: true }
+  );
   if (!selected) {
     return; // Nothing to do
   }