+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.0.8
 - Add `Open Macro…` command.
 - Fix: incorrect tracking of running macros in case of error.
+- Open-macro dropdown is configurable.
 
 ## 0.0.7
 - `[Run|Debug] Macro…` commands take in an optional `pathOrUri` argument.